Move the REST API registration into a new class
There will be more to add,
so this should keep it better organized.
