+++ php/admin/migration/class-submenu.php
@@ -9,7 +9,6 @@
 
 namespace Block_Lab\Admin\Migration;
 
-use WP_REST_Response;
 use Block_Lab\Component_Abstract;
 
 /**
@@ -52,8 +51,6 @@ class Submenu extends Component_Abstract {
 		add_action( 'admin_menu', [ $this, 'add_submenu_page' ], 9 );
 		add_action( 'admin_enqueue_scripts', [ $this, 'enqueue_scripts' ] );
 		add_action( 'admin_bar_init', [ $this, 'maybe_activate_plugin' ] );
-		add_action( 'rest_api_init', [ $this, 'register_route_migrate_post_content' ] );
-		add_action( 'rest_api_init', [ $this, 'register_route_migrate_post_type' ] );
 	}
 
 	/**
@@ -176,57 +173,5 @@ class Submenu extends Component_Abstract {
 				admin_url( 'plugins.php' )
 			)
 		);
-	}
-
-	/**
-	 * Registers a route to migrate the post content to the new namespace.
-	 */
-	public function register_route_migrate_post_content() {
-		register_rest_route(
-			block_lab()->get_slug(),
-			'migrate-post-content',
-			[
-				'methods'             => 'POST',
-				'callback'            => [ $this, 'get_migrate_post_content_response' ],
-				'permission_callback' => function() {
-					return current_user_can( self::MIGRATION_CAPABILITY );
-				},
-			]
-		);
-	}
-
-	/**
-	 * Gets the REST API response for the post content migration.
-	 *
-	 * @return WP_REST_Response The response to the request.
-	 */
-	public function get_migrate_post_content_response() {
-		return rest_ensure_response( ( new Post_Content( 'block-lab', 'genesis-custom-blocks' ) )->migrate_all() );
-	}
-
-	/**
-	 * Registers a route to migrate the post type.
-	 */
-	public function register_route_migrate_post_type() {
-		register_rest_route(
-			block_lab()->get_slug(),
-			'migrate-post-type',
-			[
-				'methods'             => 'POST',
-				'callback'            => [ $this, 'get_migrate_post_type_response' ],
-				'permission_callback' => function() {
-					return current_user_can( self::MIGRATION_CAPABILITY );
-				},
-			]
-		);
-	}
-
-	/**
-	 * Gets the REST API response for the post type migration.
-	 *
-	 * @return WP_REST_Response The response to the request.
-	 */
-	public function get_migrate_post_type_response() {
-		return rest_ensure_response( ( new Post_Type( 'block_lab', 'block-lab', 'block_lab', 'genesis_custom_block', 'genesis-custom-blocks', 'genesis_custom_blocks' ) )->migrate_all() );
 	}
 }
